+++ realm/realm-library/src/main/java/io/realm/RealmList.java
@@ -18,10 +18,11 @@ package io.realm;
 
 import java.util.AbstractList;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
+import java.util.Date;
 import java.util.List;
 
-import io.realm.exceptions.RealmException;
 import io.realm.internal.InvalidRow;
 import io.realm.internal.LinkView;
 
@@ -43,7 +44,7 @@ import io.realm.internal.LinkView;
  * @param <E> the class of objects in list.
  */
 
-public class RealmList<E extends RealmObject> extends AbstractList<E> {
+public class RealmList<E extends RealmObject> extends AbstractList<E> implements OrderedRealmCollection<E> {
 
     private static final String ONLY_IN_MANAGED_MODE_MESSAGE = "This method is only available in managed mode";
     private static final String NULL_OBJECTS_NOT_ALLOWED_MESSAGE = "RealmList does not accept null values";
@@ -117,8 +118,18 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> {
         if (!managedMode) {
             return false;
         }
-
+        if (realm == null || realm.isClosed()) {
+            return false;
+        }
         return isAttached();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public BaseRealm getRealm() {
+        return realm;
     }
 
     private boolean isAttached() {
@@ -149,6 +160,9 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> {
         checkValidObject(object);
         if (managedMode) {
             checkValidView();
+            if (location < 0 || location > size()) {
+                throw new IndexOutOfBoundsException("Invalid index " + location + ", size is " + size());
+            }
             object = copyToRealmIfNeeded(object);
             view.insert(location, object.row.getIndex());
         } else {
@@ -303,23 +317,102 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> {
     }
 
     /**
+     * Removes one instance of the specified object from this {@code Collection} if one
+     * is contained . This implementation iterates over this
+     * {@code Collection} and tests for each element {@code e} returned by the iterator,
+     * whether {@code e} is equal to the given object. If {@code object != null}
+     * then this test is performed using {@code object.equals(e)}, otherwise
+     * using {@code object == null}. If an element equal to the given object is
+     * found, then the {@code remove} method is called on the iterator and
+     * {@code true} is returned, {@code false} otherwise. If the iterator does
+     * not support removing elements, an {@code UnsupportedOperationException}
+     * is thrown.
+     *
+     * @param object the object to remove.
+     * @return {@code true} if this {@code Collection} is modified, {@code false} otherwise.
+     * @throws ClassCastException if the object passed is not of the correct type.
+     * @throws NullPointerException  if {@code object} is {@code null}.
+     */
+    @Override
+    public boolean remove(Object object) {
+        if (managedMode && !realm.isInTransaction()) {
+            throw new IllegalStateException("Objects can only be removed from inside a write transaction");
+        }
+        return super.remove(object);
+    }
+
+    /**
+     * Removes all occurrences in this {@code Collection} of each object in the
+     * specified {@code Collection}. After this method returns none of the
+     * elements in the passed {@code Collection} can be found in this {@code Collection}
+     * anymore.
+     * <p>
+     * This implementation iterates over this {@code Collection} and tests for each
+     * element {@code e} returned by the iterator, whether it is contained in
+     * the specified {@code Collection}. If this test is positive, then the {@code
+     * remove} method is called on the iterator.
+     *
+     * @param collection the collection of objects to remove.
+     * @return {@code true} if this {@code Collection} is modified, {@code false} otherwise.
+     * @throws ClassCastException if one or more elements of {@code collection} isn't of the correct type.
+     * @throws NullPointerException if {@code collection} contains at least one {@code null} element.
+     * @throws NullPointerException if {@code collection} is {@code null}.
+     */
+    @Override
+    public boolean removeAll(Collection<?> collection) {
+        if (managedMode && !realm.isInTransaction()) {
+            throw new IllegalStateException("Objects can only be removed from inside a write transaction");
+        }
+        return super.removeAll(collection);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean deleteFirstFromRealm() {
+        if (managedMode) {
+            if (size() > 0) {
+                deleteFromRealm(0);
+                return true;
+            } else {
+                return false;
+            }
+        } else {
+            throw new UnsupportedOperationException("deleteFirstFromRealm() not supported by un-managed RealmLists.");
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean deleteLastFromRealm() {
+        if (managedMode) {
+            if (size() > 0) {
+                deleteFromRealm(size() - 1);
+                return true;
+            } else {
+                return false;
+            }
+        } else {
+            throw new UnsupportedOperationException("deleteLastFromRealm() not supported by un-managed RealmLists.");
+        }
+    }
+
+    /**
      * Removes all elements from this list and delete them from the corresponding Realm. This method can be called on a
      * non-managed {@link RealmList} if all of the RealmObjects in the list are managed by Realm.
+     *
+     * Use {@link #deleteAllFromRealm()} instead.
      *
      * @throws IllegalStateException if the Realm instance has been closed, the parent object has been removed, the
      * method is called in a wrong thread or any RealmObject in the list is not managed by Realm.
      * @see #clear()
      */
+    @Deprecated
     public void removeAllFromRealm() {
-        if (managedMode) {
-            checkValidView();
-            view.removeAllTargetRows();
-        } else {
-            for (RealmObject object : nonManagedList) {
-                object.removeFromRealm();
-            }
-            nonManagedList.clear();
-        }
+        deleteAllFromRealm();
     }
 
     /**
@@ -374,6 +467,59 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> {
     }
 
     /**
+     * {@inheritDoc}
+     */
+    @Override
+    public RealmResults<E> sort(String fieldName) {
+        return this.sort(fieldName, Sort.ASCENDING);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public RealmResults<E> sort(String fieldName, Sort sortOrder) {
+        if (managedMode) {
+            return this.where().findAllSorted(fieldName, sortOrder);
+        } else {
+            throw new UnsupportedOperationException("sort() not supported by un-managed RealmList.");
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public RealmResults<E> sort(String fieldName1, Sort sortOrder1, String fieldName2, Sort sortOrder2) {
+        return sort(new String[]{fieldName1, fieldName2}, new Sort[]{sortOrder1, sortOrder2});
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public RealmResults<E> sort(String[] fieldNames, Sort[] sortOrders) {
+        if (managedMode) {
+            return where().findAllSorted(fieldNames, sortOrders);
+        } else {
+            throw new UnsupportedOperationException("sort() not supported by un-managed RealmList.");
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void deleteFromRealm(int location) {
+        if (managedMode) {
+            checkValidView();
+            view.removeTargetRow(location);
+        } else {
+            throw new UnsupportedOperationException("deleteFromRealm() not supported by un-managed RealmLists.");
+        }
+    }
+
+    /**
      * Returns the number of elements in this {@code List}.
      *
      * @return the number of elements in this {@code List}.
@@ -402,8 +548,114 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> {
             checkValidView();
             return RealmQuery.createQueryFromList(this);
         } else {
-            throw new RealmException(ONLY_IN_MANAGED_MODE_MESSAGE);
+            throw new UnsupportedOperationException(ONLY_IN_MANAGED_MODE_MESSAGE);
         }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Number min(String fieldName) {
+        if (managedMode) {
+            return this.where().min(fieldName);
+        } else {
+            throw new UnsupportedOperationException("min() not supported by un-managed RealmLists.");
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Number max(String fieldName) {
+        if (managedMode) {
+            return this.where().max(fieldName);
+        } else {
+            throw new UnsupportedOperationException("max() not supported by un-managed RealmLists.");
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Number sum(String fieldName) {
+        if (managedMode) {
+            return this.where().sum(fieldName);
+        } else {
+            throw new UnsupportedOperationException("sum() not supported by un-managed RealmLists.");
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public double average(String fieldName) {
+        if (managedMode) {
+            return this.where().average(fieldName);
+        } else {
+            throw new UnsupportedOperationException("average() not supported by un-managed RealmLists.");
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Date maxDate(String fieldName) {
+        if (managedMode) {
+            return this.where().maximumDate(fieldName);
+        } else {
+            throw new UnsupportedOperationException("maxDate() not supported by un-managed RealmLists.");
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Date minDate(String fieldName) {
+        if (managedMode) {
+            return this.where().minimumDate(fieldName);
+        } else {
+            throw new UnsupportedOperationException("maxDate() not supported by un-managed RealmLists.");
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean deleteAllFromRealm() {
+        if (managedMode) {
+            checkValidView();
+            if (size() > 0) {
+                view.removeAllTargetRows();
+                return true;
+            } else {
+                return false;
+            }
+        } else {
+            throw new UnsupportedOperationException("deleteAllFromRealm() not supported by un-managed RealmLists.");
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean isLoaded() {
+        return true; // Managed RealmLists are always loaded, Un-managed RealmLists return true pr. the contract.
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean load() {
+        return true; // Managed RealmLists are always loaded, Un-managed RealmLists return true pr. the contract.
     }
 
     /**
@@ -420,6 +672,7 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> {
     public boolean contains(Object object) {
         boolean contains = false;
         if (managedMode) {
+            realm.checkIfValid();
             if (object instanceof RealmObject) {
                 RealmObject realmObject = (RealmObject) object;
                 if (realmObject.row != null && realm.getPath().equals(realmObject.realm.getPath()) && realmObject.row != InvalidRow.INSTANCE) {
@@ -448,7 +701,7 @@ public class RealmList<E extends RealmObject> extends AbstractList<E> {
     private void checkValidView() {
         realm.checkIfValid();
         if (view == null || !view.isAttached()) {
-            throw new IllegalStateException("Realm instance has been closed or parent object has been removed.");
+            throw new IllegalStateException("Realm instance has been closed or this object or its parent has been deleted.");
         }
     }